% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrc_cols.R
\name{mrc_cols}
\alias{mrc_cols}
\title{Function to extract mrc_cols colors as hex codes}
\usage{
mrc_cols(...)
}
\arguments{
\item{...}{Character names of mrc_cols}
}
\description{
Function to extract mrc_cols colors as hex codes
}
