% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrc_pal.R
\name{mrc_pal}
\alias{mrc_pal}
\title{Return function to interpolate a mrc color palette}
\usage{
mrc_pal(palette = "primary", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a mrc color palette
}
